Holy macaroni - Huge commit
This does the bulk (I think) of the work for changing
LandmarkGroup to being a shape. A lot of files are touched
because the .lms property (which has been deprecated) was being
used all over the shop.

As discussed with @jabooth and @nontas, this essentially does
the following:

  - LandmarkGroup is now a PointUndirectedGraph with special
    labelling abilities.
  - LandmarkGroup is no longer a MutableMapping and now has
    explicit methods for getting/setting labels
  - LandmarkGroup label mutation maintains immutability - so
    whenever labels are added or removed a new object is returned.
  - .lms and .n_landmarks are deprecated but maintained for
    compatibility
  - A dummy LandmarkGroup is maintained in menpo.landmark for
    pickle compatibility - unpickling old-style LandmarkGroup
    classes will raise a warning and actually return a new
    style class.
  - All labels have been updated for the new class.

TODO:
  - Thoroughly check visualization - this includes reviewing
    the function of the 'view_landmarks' method.
  - Make sure the new unpickling is actually unit tested
  - Fix all the widgets
