+++ menpo/image/test/image_copy_test.py
@@ -13,8 +13,8 @@ def test_image_copy():
     im_copy = im.copy()
 
     assert (not is_same_array(im.pixels, im_copy.pixels))
-    assert (not is_same_array(im_copy.landmarks['test'].lms.points,
-                              im.landmarks['test'].lms.points))
+    assert (not is_same_array(im_copy.landmarks['test'].points,
+                              im.landmarks['test'].points))
 
 
 def test_booleanimage_copy():
@@ -25,8 +25,8 @@ def test_booleanimage_copy():
     im_copy = im.copy()
 
     assert (not is_same_array(im.pixels, im_copy.pixels))
-    assert (not is_same_array(im_copy.landmarks['test'].lms.points,
-                              im.landmarks['test'].lms.points))
+    assert (not is_same_array(im_copy.landmarks['test'].points,
+                              im.landmarks['test'].points))
 
 
 def test_maskedimage_copy():
@@ -37,5 +37,5 @@ def test_maskedimage_copy():
     im_copy = im.copy()
 
     assert (not is_same_array(im.pixels, im_copy.pixels))
-    assert (not is_same_array(im_copy.landmarks['test'].lms.points,
-                              im.landmarks['test'].lms.points))
+    assert (not is_same_array(im_copy.landmarks['test'].points,
+                              im.landmarks['test'].points))
