+++ menpo/image/test/image_warp_test.py
@@ -138,8 +138,8 @@ def test_rescale_return_transform():
     img_back = cropped_img.warp_to_shape(img.shape, transform.pseudoinverse())
     assert_allclose(img_back.shape, img.shape)
     assert_allclose(img_back.pixels, img.pixels)
-    assert_allclose(img_back.landmarks['test'].lms.points,
-                    img.landmarks['test'].lms.points)
+    assert_allclose(img_back.landmarks['test'].points,
+                    img.landmarks['test'].points)
 
 
 def test_sample_image():
@@ -223,7 +223,7 @@ def test_mirror_horizontal_image():
                     np.array([[[9., 10., 11., 12.],
                                [5., 6., 7., 8.],
                                [1., 2., 3., 4.]]]))
-    assert_allclose(mirrored_img.landmarks['temp'].lms.points,
+    assert_allclose(mirrored_img.landmarks['temp'].points,
                     np.array([[1., 1.], [1., 2.], [0., 1.], [0., 2.]]))
 
 
@@ -238,7 +238,7 @@ def test_mirror_vertical_image():
                     np.array([[[4., 3., 2., 1.],
                                [8., 7., 6., 5.],
                                [12., 11., 10., 9.]]]))
-    assert_allclose(mirrored_img.landmarks['temp'].lms.points,
+    assert_allclose(mirrored_img.landmarks['temp'].points,
                     np.array([[1., 3.], [1., 2.], [2., 2.], [2., 1.]]))
 
 
@@ -260,8 +260,8 @@ def test_mirror_return_transform():
     img_back = cropped_img.warp_to_shape(img.shape, transform.pseudoinverse())
     assert_allclose(img_back.shape, img.shape)
     assert_allclose(img_back.pixels, img.pixels)
-    assert_allclose(img_back.landmarks['test'].lms.points,
-                    img.landmarks['test'].lms.points)
+    assert_allclose(img_back.landmarks['test'].points,
+                    img.landmarks['test'].points)
 
 
 def test_rotate_image_90_180():
@@ -277,7 +277,7 @@ def test_rotate_image_90_180():
                                                    [3., 7., 11.],
                                                    [2., 6., 10.],
                                                    [1., 5., 9.]]]))
-    assert_almost_equal(rotated_img.landmarks['temp'].lms.points,
+    assert_almost_equal(rotated_img.landmarks['temp'].points,
                         np.array([[2., 1.], [1., 1.], [2., 2.], [1., 2.]]))
 
     # rotate 180 degrees
@@ -286,7 +286,7 @@ def test_rotate_image_90_180():
     assert_allclose(rotated_img.pixels, np.array([[[12., 11., 10., 9.],
                                                    [8., 7., 6., 5.],
                                                    [4., 3., 2., 1.]]]))
-    assert_almost_equal(rotated_img.landmarks['temp'].lms.points,
+    assert_almost_equal(rotated_img.landmarks['temp'].points,
                         np.array([[1., 2.], [1., 1.], [0., 2.], [0., 1.]]))
 
 
@@ -304,7 +304,7 @@ def test_rotate_image_45():
                                [1., 6., 7., 11., 16.],
                                [0., 5., 10., 15., 15.],
                                [0., 0., 13., 14., 0.]]]))
-    assert_almost_equal(rotated_img.landmarks['temp'].lms.points,
+    assert_almost_equal(rotated_img.landmarks['temp'].points,
                         np.array([[2.121, 1.414], [1.414, 2.121],
                                   [2.828, 2.121], [2.121, 2.828]]), decimal=3)
 
@@ -317,8 +317,8 @@ def test_rotate_return_transform():
     img_back = cropped_img.warp_to_shape(img.shape, transform.pseudoinverse())
     assert_allclose(img_back.shape, img.shape)
     assert_allclose(img_back.pixels, img.pixels)
-    assert_allclose(img_back.landmarks['test'].lms.points,
-                    img.landmarks['test'].lms.points)
+    assert_allclose(img_back.landmarks['test'].points,
+                    img.landmarks['test'].points)
 
 
 def test_maskedimage_retain_shape():
