+++ menpo/landmark/__init__.py
@@ -1,3 +1,3 @@
-from .base import LandmarkGroup, LandmarkManager, Landmarkable
+from .base import LandmarkManager, Landmarkable
 from .exceptions import LabellingError
 from .labels import *
