+++ menpo/landmark/labels/human/pose.py
@@ -54,7 +54,7 @@ def pose_lsp_14_to_pose_lsp_14(pcloud):
     ----------
     .. [1] http://www.comp.leeds.ac.uk/mat4saj/lsp.html
     """
-    from menpo.shape import PointUndirectedGraph
+    from menpo.shape import LandmarkGroup
 
     n_expected_points = 14
     validate_input(pcloud, n_expected_points)
@@ -77,15 +77,15 @@ def pose_lsp_14_to_pose_lsp_14(pcloud):
         head_connectivity
     ])
 
-    new_pcloud = PointUndirectedGraph.init_from_edges(pcloud.points,
-                                                      all_connectivity)
-
     mapping = OrderedDict()
     mapping['left_leg'] = left_leg_indices
     mapping['right_leg'] = right_leg_indices
     mapping['left_arm'] = left_arm_indices
     mapping['right_arm'] = right_arm_indices
     mapping['head'] = head_indices
+
+    new_pcloud = LandmarkGroup.init_from_indices_mapping(
+        pcloud.points, all_connectivity, mapping)
 
     return new_pcloud, mapping
 
@@ -140,7 +140,7 @@ def pose_human36M_32_to_pose_human36M_32(pcloud):
     ----------
     .. [1] http://vision.imar.ro/human3.6m/
     """
-    from menpo.shape import PointUndirectedGraph
+    from menpo.shape import LandmarkGroup
 
     n_expected_points = 32
     validate_input(pcloud, n_expected_points)
@@ -175,9 +175,6 @@ def pose_human36M_32_to_pose_human36M_32(pcloud):
         right_hand_connectivity, torso_connectivity
     ])
 
-    new_pcloud = PointUndirectedGraph.init_from_edges(pcloud.points,
-                                                      all_connectivity)
-
     mapping = OrderedDict()
     mapping['pelvis'] = pelvis_indices
     mapping['right_leg'] = right_leg_indices
@@ -189,6 +186,9 @@ def pose_human36M_32_to_pose_human36M_32(pcloud):
     mapping['right_arm'] = right_arm_indices
     mapping['right_hand'] = right_hand_indices
     mapping['torso'] = torso_indices
+
+    new_pcloud = LandmarkGroup.init_from_indices_mapping(
+        pcloud.points, all_connectivity, mapping)
 
     return new_pcloud, mapping
 
@@ -215,7 +215,7 @@ def pose_human36M_32_to_pose_human36M_17(pcloud):
     ----------
     .. [1] http://vision.imar.ro/human3.6m/
     """
-    from menpo.shape import PointUndirectedGraph
+    from menpo.shape import LandmarkGroup
 
     n_expected_points = 32
     validate_input(pcloud, n_expected_points)
@@ -245,12 +245,6 @@ def pose_human36M_32_to_pose_human36M_17(pcloud):
         right_arm_connectivity, torso_connectivity
     ])
 
-    # Ignore duplicate points, sole and palms
-    ind = np.hstack([np.arange(0, 4), np.arange(6, 9), np.arange(12, 16),
-                     np.arange(17, 20), np.arange(25, 28)])
-    new_pcloud = PointUndirectedGraph.init_from_edges(
-        pcloud.points[ind], all_connectivity)
-
     mapping = OrderedDict()
     mapping['pelvis'] = pelvis_indices
     mapping['right_leg'] = right_leg_indices
@@ -260,5 +254,11 @@ def pose_human36M_32_to_pose_human36M_17(pcloud):
     mapping['left_arm'] = left_arm_indices
     mapping['right_arm'] = right_arm_indices
     mapping['torso'] = torso_indices
+
+    # Ignore duplicate points, sole and palms
+    ind = np.hstack([np.arange(0, 4), np.arange(6, 9), np.arange(12, 16),
+                     np.arange(17, 20), np.arange(25, 28)])
+    new_pcloud = LandmarkGroup.init_from_indices_mapping(
+        pcloud.points[ind], all_connectivity, mapping)
 
     return new_pcloud, mapping
