+++ menpo/shape/mesh/textured.py
@@ -176,7 +176,7 @@ class TexturedTriMesh(TriMesh):
                  axes_font_style='normal', axes_font_weight='normal',
                  axes_x_limits=None, axes_y_limits=None, axes_x_ticks=None,
                  axes_y_ticks=None, figure_size=(10, 8),
-                 label=None):
+                 label=None, **kwargs):
         r"""
         Visualization of the TriMesh in 2D. Currently, explicit textured TriMesh
         viewing is not supported, and therefore viewing falls back to untextured
