+++ menpo/shape/test/pointcloud_test.py
@@ -69,8 +69,8 @@ def test_pointcloud_copy_method():
     p_copy = p.copy()
 
     assert (not is_same_array(p_copy.points, p.points))
-    assert (not is_same_array(p_copy.landmarks['test'].lms.points,
-                              p.landmarks['test'].lms.points))
+    assert (not is_same_array(p_copy.landmarks['test'].points,
+                              p.landmarks['test'].points))
 
 
 def test_pointcloud_copy_false():
