+++ menpo/visualize/viewmatplotlib.py
@@ -515,12 +515,10 @@ class MatplotlibPointGraphViewer2d(MatplotlibRenderer):
 
 
 class MatplotlibLandmarkViewer2d(MatplotlibRenderer):
-    def __init__(self, figure_id, new_figure, group, pointcloud,
-                 labels_to_masks):
+    def __init__(self, figure_id, new_figure, group, landmark_group):
         super(MatplotlibLandmarkViewer2d, self).__init__(figure_id, new_figure)
         self.group = group
-        self.pointcloud = pointcloud
-        self.labels_to_masks = labels_to_masks
+        self.landmark_group = landmark_group
 
     def render(self, image_view=False, render_lines=True, line_colour='r',
                line_style='-', line_width=1, render_markers=True,
@@ -545,15 +543,13 @@ class MatplotlibLandmarkViewer2d(MatplotlibRenderer):
                axes_x_limits=None, axes_y_limits=None, axes_x_ticks=None,
                axes_y_ticks=None, figure_size=(10, 8)):
         import matplotlib.lines as mlines
-        from menpo.shape import TriMesh
-        from menpo.shape.graph import PointGraph
         import matplotlib.pyplot as plt
 
         # Regarding the labels colours, we may get passed either no colours (in
         # which case we generate random colours) or a single colour to colour
         # all the labels with
         # TODO: All marker and line options could be defined as lists...
-        n_labels = len(self.labels_to_masks)
+        n_labels = self.landmark_group.n_labels
         line_colour = _check_colours_list(
             render_lines, line_colour, n_labels,
             'Must pass a list of line colours with length n_labels or a single '
@@ -568,12 +564,13 @@ class MatplotlibLandmarkViewer2d(MatplotlibRenderer):
             'a single marker edge colour for all labels.')
 
         # check axes limits
+        points = self.landmark_group.points
         if image_view:
-            min_y, min_x = np.min(self.pointcloud.points, axis=0)
-            max_y, max_x = np.max(self.pointcloud.points, axis=0)
+            min_y, min_x = np.min(points, axis=0)
+            max_y, max_x = np.max(points, axis=0)
         else:
-            min_x, min_y = np.min(self.pointcloud.points, axis=0)
-            max_x, max_y = np.max(self.pointcloud.points, axis=0)
+            min_x, min_y = np.min(points, axis=0)
+            max_x, max_y = np.max(points, axis=0)
         axes_x_limits, axes_y_limits = _parse_axes_limits(
             min_x, max_x, min_y, max_y, axes_x_limits, axes_y_limits)
 
@@ -612,10 +609,7 @@ class MatplotlibLandmarkViewer2d(MatplotlibRenderer):
 
             # set legend entry
             if render_legend:
-                tmp_line = 'None'
-                if (render_lines and
-                        (isinstance(pc, PointGraph) or isinstance(pc, TriMesh))):
-                    tmp_line = line_style
+                tmp_line = line_style
                 tmp_marker = marker_style if render_markers else 'None'
                 legend_handles.append(
                     mlines.Line2D([], [], linewidth=line_width,
@@ -650,11 +644,8 @@ class MatplotlibLandmarkViewer2d(MatplotlibRenderer):
         return self
 
     def _build_sub_pointclouds(self):
-        sub_pointclouds = []
-        for label, indices in self.labels_to_masks.items():
-            mask = self.labels_to_masks[label]
-            sub_pointclouds.append((label, self.pointcloud.from_mask(mask)))
-        return sub_pointclouds
+        return [(label, self.landmark_group.get_label(label))
+                for label in self.landmark_group.labels]
 
 
 class MatplotlibAlignmentViewer2d(MatplotlibRenderer):
